chore: moved projectEdge-file to lib
