+++ src/lib/projection/projectionEdgeQuality.ts
